Create a module to give NightReader some methods from NightWriter
